moved roles.js out fo seeds to models
